modified requestService
