+++ client/src/componentes/RequestService/RequestService.jsx
@@ -3,35 +3,45 @@ import React from "react";
 import { useState } from "react";
 import { Link, useParams, useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { getDetail, getUserEmail, postRequest } from "../../redux/actions/index.js";
+import {
+  getDetail,
+  getUserEmail,
+  postRequest,
+} from "../../redux/actions/index.js";
 import { useEffect } from "react";
 import { useAuth } from "../../context/authContext";
 import { Box, Typography, Button, TextField } from "@mui/material";
-import userImg from "../../../src/user.png";
-import Navbar from "../PrivateRoute/Navbar.jsx";
+import userImg from "../../src/user.png";
+import Navbar from "../PrivateRoute/Navbar";
+
 
 export default function RequestService(props) {
-
   const { user } = useAuth();
+  console.log(useAuth())
 
   const [request, setRequest] = useState({
     day: "",
     hours: "",
   });
+
+  const [loading, setLoading] = useState(true)
+
   const navigate = useNavigate();
   const dispatch = useDispatch();
   const { id } = useParams();
 
+  const service = useSelector((state) => state.serviceDetail);
+  const userDb = useSelector(state => state.filter)
+
   useEffect(() => {
     dispatch(getDetail(id));
     dispatch(getUserEmail(user?.email))
-  }, [dispatch]);
+    setLoading(false)
+  }, [dispatch, user?.email]);
   console.log(id, 'ID DE SERVICE');
 
+  const weekDays = ['Lunes','Martes','Miércoles','Jueves','Viernes','Sábado','Domingo'];
 
-  const service = useSelector((state) => state.serviceDetail);
-  const userDb = useSelector((state) => state.filter)
-  
 
   const handleOnChange = (e) => {
     e.preventDefault();
@@ -42,23 +52,28 @@ export default function RequestService(props) {
   };
 
   const handleDay = (e) => {
-    if (!request.day.includes(e.target.value)) {
-      setRequest({
-        ...request,
-        day: [...request.day, e.target.value],
-      });
+    if(request.day !== ''){
+      document.getElementById(request.day).style = 'color: #1F2937'
     }
+    e.target.style.cssText = 'color: white; background-color: #1F2937';
+    setRequest({
+      ...request,
+      day: e.target.value
+    });
   };
 
   const handleSubmit = (e) => {
     request.day = request.day.join(",");
     e.preventDefault();
-    let requestService = { ...request, service_id: service.id, requester_id: userDb[0].id };
+    let requestService = {
+      ...request,
+      service_id: service.id,
+      requester_id: userDb[0].id,
+    };
     dispatch(postRequest(requestService));
     setRequest({
       day: "",
       hours: "",
-
     });
     navigate("/home");
   };
@@ -67,28 +82,35 @@ export default function RequestService(props) {
     container: {
       display: "flex",
       alignItems: "center",
-      justifyContent: "center",
+      justifyContent: 'center',
       height: "100vh",
       width: "100vw",
       backgroundColor: "#E5E7EB",
       color: "#1F2937",
     },
     containerRequest: {
-      width: "50%",
+      width: "60%",
+      margin: "20px 10px 20px 20px"
     },
     containerUser: {
+      margin: "20px 20px 20px 10px",
+      flexDirection: 'column',
+      width: "40%",
       display: "flex",
+    },
+    userDetail: {
+      display: 'flex',
+      flexDirection: 'row',
       justifyContent: "space-between",
-      alignItems: "center",
+      width: '100%',
       border: "solid black 2px",
-      padding: "20px",
     },
     containerService: {
       display: "flex",
+      flexDirection:'column',
       justifyContent: "center",
       alignItems: "center",
       border: "solid black 2px",
-      marginTop: "20px",
       padding: "20px",
     },
     containerRequestForm: {
@@ -103,13 +125,33 @@ export default function RequestService(props) {
       display: "flex",
     },
     userPic: {
-      width: "50px",
+      width: "100px",
+      borderRadius: "50%",
+      padding: '20px'
     },
+    userName: {
+      display:'flex',
+      flexDirection: 'column',
+      justifyContent:'center',
+      padding: '20px'
+    },
+    reviews: {
+      width: "100%",
+      border: "solid black 2px",
+      marginTop: '20px'
+    },
+    selectedButton: {
+      color: 'white',
+      backgroundColor: 'black'
+    }
   };
 
-  return (
+  console.log(userDb)
+
+  if(loading) return <h1>loading</h1>
+  else return (
     <div>
-      <Navbar/>
+    <Navbar/>
       {
         user?.email === null ? <p>
         No tienes acceso a estos datos ya que ingresaste como un usuario
@@ -117,52 +159,51 @@ export default function RequestService(props) {
         servicios.
         <Link to="/register">Registrarse</Link>
       </p>
-        : <Box style={styles.container}>
+        : 
+    <Box style={styles.container}>
+      <Box sx={{display:'flex', width: '100%', margin: '20px'}}>
         <Box style={styles.containerRequest}>
-          <Box style={styles.containerUser}>
-            <Typography sx={{ textAlign: "center" }} variant="h4">
-              Acá va el usuario
-            </Typography>
-            <img style={styles.userPic} src={userImg} alt="" />
-          </Box>
           <Box style={styles.containerService}>
-            <Typography sx={{ textAlign: "center" }} variant="h4">
-              {service.name}
-            </Typography>
-            <Box style={styles.box}>
-              <Typography variant="h7">Description: </Typography>
-              <Typography variant="h7">{service.description}</Typography>
+            <Box sx={{ display: 'flex', justifyContent: 'space-between', width: '100%'}}>
+              <Typography variant="h4">
+                {service.name}
+              </Typography>
+              <Box style={styles.box}>
+                <Typography variant="h4">Precio: </Typography>
+                <Typography
+                  sx={{color:'green', marginLeft:'10px'}} 
+                  variant="h4">
+                  {` $${service.price}`} 
+                </Typography>
+              </Box>
             </Box>
-            <Box style={styles.box}>
-              <Typography variant="h7">{`Price: $${service.price}`} </Typography>
-            </Box>
-            <Box style={styles.box}>
-              <Typography variant="h7">{`Rating: ${service.rating}`} </Typography>
+            <Box sx={{width:'100%', padding: '20px'}}>
+              <Typography variant="h5">{`Descripción: ${service.description}`}</Typography>
             </Box>
           </Box>
           <Box style={styles.containerRequestForm}>
             <form onSubmit={(e) => handleSubmit(e)}>
-              {/* <Box style={styles.box}>
-                <TextField
-                  id="outlined-basic"
-                  label="día"
-                  variant="outlined"
-                  style={styles.input}
-                  type="text"
-                  name="day"
-                  value={request.day}
-                  onChange={handleOnChange}
-                />
-              </Box> */}
-  
-              {service.day?.split(",").map((el) => {
-                return (
-                  <Button value={el} onClick={(e) => handleDay(e)}>
-                    {el}
+              <Box sx={{display:'flex', justifyContent:'center'}}>
+                {
+                weekDays.map((el) => {
+                  if(service.day?.split(',').includes(el)){
+                    return <Button 
+                      variant="outlined"
+                      id={el} 
+                      value={el} 
+                      onClick={(e) => handleDay(e)}
+                      sx={{color: "#1F2937", margin: '5px'}}
+                    >
+                      {el}
+                    </Button>
+                  }else return <Button 
+                    disabled variant="outlined"
+                    sx={{color: "#1F2937", margin: '5px'}}
+                  >
+                  {el}
                   </Button>
-                );
-              })}
-  
+                })}
+              </Box>
               <Box style={styles.box}>
                 <TextField
                   id="outlined-basic"
@@ -175,13 +216,13 @@ export default function RequestService(props) {
                   onChange={handleOnChange}
                 />
               </Box>
-  
+
               <Box sx={{ display: "flex", justifyContent: "space-around" }}>
-                <Button>
-                  <Link style={{ textDecoration: "none" }} to="/home">
-                    <label style={{ color: "#1F2937" }}>Volver atras</label>
-                  </Link>
-                </Button>
+                <Link style={{ textDecoration: "none" }} to="/home">
+                  <Button style={{ color: "#1F2937" }}>
+                    Volver atras
+                  </Button>
+                </Link>
                 <Button sx={{ color: "#1F2937" }} type="submit">
                   Solicitar
                 </Button>
@@ -189,8 +230,31 @@ export default function RequestService(props) {
             </form>
           </Box>
         </Box>
+        <Box style={styles.containerUser}>
+          <Box style={styles.userDetail}>
+            <Box style={styles.userName}>
+              <Typography variant="h4">
+                {userDb[0]?.firstName}
+              </Typography>
+              <Typography variant="h6">
+                {userDb[0]?.lastName}
+              </Typography>
+            </Box>
+            <img
+              style={styles.userPic}
+              src={userDb[0]?.img ? userDb[0].img : userImg}
+              alt="user-pic"
+            />
+          </Box>
+          <Box style={styles.reviews}>
+              reviews
+          </Box>
+
+        </Box>
       </Box>
+    </Box>
       }
-    </div>
+  </div>
   );
+
 }
